+++ server/BudgetBoard.Service/Helpers/WidgetSettingsHelpers.cs
@@ -1,4 +1,5 @@
 using BudgetBoard.Service.Models;
+using BudgetBoard.Service.Models.Widgets.NetWorthWidget;
 
 namespace BudgetBoard.Service.Helpers;
