+++ server/BudgetBoard.Service/Models/Widgets/NetWorthWidget/NetWorthWidgetConfiguration.cs
@@ -1,0 +1,6 @@
+namespace BudgetBoard.Service.Models.Widgets.NetWorthWidget;
+
+public class NetWorthWidgetConfiguration
+{
+    public IEnumerable<NetWorthWidgetGroup> Groups { get; set; } = [];
+}
