+++ server/BudgetBoard.Service/Models/Widgets/NetWorthWidget/NetWorthWidgetGroup.cs
@@ -1,4 +1,11 @@
-﻿namespace BudgetBoard.Service.Models;
+﻿namespace BudgetBoard.Service.Models.Widgets.NetWorthWidget;
+
+public class NetWorthWidgetGroup
+{
+    public Guid ID { get; set; } = Guid.NewGuid();
+    public int Index { get; set; } = 0;
+    public IEnumerable<NetWorthWidgetLine> Lines { get; set; } = [];
+}
 
 public interface INetWorthWidgetGroupReorderRequest
 {
