+++ server/BudgetBoard.Service/Models/Widgets/NetWorthWidget/NetWorthWidgetLine.cs
@@ -1,4 +1,12 @@
-﻿namespace BudgetBoard.Service.Models;
+﻿namespace BudgetBoard.Service.Models.Widgets.NetWorthWidget;
+
+public class NetWorthWidgetLine
+{
+    public Guid ID { get; set; } = Guid.NewGuid();
+    public string Name { get; set; } = string.Empty;
+    public List<NetWorthWidgetCategory> Categories { get; set; } = [];
+    public int Index { get; set; } = 0;
+}
 
 public interface INetWorthWidgetLineCreateRequest
 {
