+++ server/BudgetBoard.Service/NetWorthWidgetGroupService.cs
@@ -3,6 +3,7 @@ using BudgetBoard.Database.Data;
 using BudgetBoard.Database.Models;
 using BudgetBoard.Service.Interfaces;
 using BudgetBoard.Service.Models;
+using BudgetBoard.Service.Models.Widgets.NetWorthWidget;
 using BudgetBoard.Service.Resources;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Localization;
